+++ src/python/pants/backend/experimental/go/register.py
@@ -1,12 +1,12 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.backend.go import build, pkg, target_type_rules
+from pants.backend.go import build, target_type_rules
 from pants.backend.go import target_types as go_target_types
 from pants.backend.go.goals import custom_goals, package_binary, tailor
 from pants.backend.go.subsystems import golang
 from pants.backend.go.target_types import GoBinary, GoExternalPackageTarget, GoModule, GoPackage
-from pants.backend.go.util_rules import go_mod, import_analysis, sdk
+from pants.backend.go.util_rules import go_mod, go_pkg, import_analysis, sdk
 
 
 def target_types():
@@ -20,7 +20,7 @@ def rules():
         *go_target_types.rules(),
         *import_analysis.rules(),
         *go_mod.rules(),
-        *pkg.rules(),
+        *go_pkg.rules(),
         *sdk.rules(),
         *tailor.rules(),
         *target_type_rules.rules(),
