+++ src/python/pants/backend/go/goals/custom_goals.py
@@ -6,15 +6,15 @@ from __future__ import annotations
 import logging
 
 from pants.backend.go.build import BuildGoPackageRequest, BuiltGoPackage
-from pants.backend.go.pkg import (
+from pants.backend.go.target_types import GoModuleSources
+from pants.backend.go.util_rules.external_module import ResolveExternalGoPackageRequest
+from pants.backend.go.util_rules.go_mod import ResolvedGoModule, ResolveGoModuleRequest
+from pants.backend.go.util_rules.go_pkg import (
     ResolvedGoPackage,
     ResolveGoPackageRequest,
     is_first_party_package_target,
     is_third_party_package_target,
 )
-from pants.backend.go.target_types import GoModuleSources
-from pants.backend.go.util_rules.external_module import ResolveExternalGoPackageRequest
-from pants.backend.go.util_rules.go_mod import ResolvedGoModule, ResolveGoModuleRequest
 from pants.engine.console import Console
 from pants.engine.fs import Workspace
 from pants.engine.goal import Goal, GoalSubsystem
