+++ src/python/pants/backend/go/goals/package_binary.py
@@ -10,8 +10,11 @@ from pants.backend.go.build import (
     GatheredImports,
     GatherImportsRequest,
 )
-from pants.backend.go.pkg import is_first_party_package_target, is_third_party_package_target
 from pants.backend.go.target_types import GoBinaryMainAddress
+from pants.backend.go.util_rules.go_pkg import (
+    is_first_party_package_target,
+    is_third_party_package_target,
+)
 from pants.backend.go.util_rules.sdk import GoSdkProcess
 from pants.build_graph.address import Address, AddressInput
 from pants.core.goals.package import (
