+++ src/python/pants/backend/go/goals/package_binary_integration_test.py
@@ -7,11 +7,11 @@ from textwrap import dedent
 
 import pytest
 
-from pants.backend.go import build, pkg, target_type_rules
+from pants.backend.go import build, target_type_rules
 from pants.backend.go.goals import package_binary
 from pants.backend.go.goals.package_binary import GoBinaryFieldSet
 from pants.backend.go.target_types import GoBinary, GoModule, GoPackage
-from pants.backend.go.util_rules import external_module, go_mod, import_analysis, sdk
+from pants.backend.go.util_rules import external_module, go_mod, go_pkg, import_analysis, sdk
 from pants.build_graph.address import Address
 from pants.core.goals.package import BuiltPackage
 from pants.core.util_rules import external_tool, source_files
@@ -30,7 +30,7 @@ def rule_runner() -> RuleRunner:
             *import_analysis.rules(),
             *package_binary.rules(),
             *build.rules(),
-            *pkg.rules(),
+            *go_pkg.rules(),
             *go_mod.rules(),
             *target_type_rules.rules(),
             *external_module.rules(),
