+++ src/python/pants/backend/go/util_rules/go_pkg.py
